move singleton registration to extensions
