+++ source/server/configuration_impl.h
@@ -21,6 +21,8 @@
 #include "common/network/resolver_impl.h"
 #include "common/network/utility.h"
 
+#include "extensions/filters/common/ratelimit/ratelimit_registration.h"
+
 namespace Envoy {
 namespace Server {
 namespace Configuration {
@@ -77,22 +79,6 @@ public:
 };
 
 /**
- * RateLimitServiceConfig that wraps the proto structure so that it can be registered as a
- * singleton.
- */
-class RateLimitServiceConfig : public Singleton::Instance {
-
-public:
-  RateLimitServiceConfig(
-      const envoy::config::ratelimit::v2::RateLimitServiceConfig& ratelimit_config)
-      : config_(ratelimit_config) {}
-
-  const envoy::config::ratelimit::v2::RateLimitServiceConfig& config_;
-};
-
-typedef std::shared_ptr<RateLimitServiceConfig> RateLimitServiceConfigPtr;
-
-/**
  * Implementation of Server::Configuration::Main that reads a configuration from a JSON file.
  */
 class MainImpl : Logger::Loggable<Logger::Id::config>, public Main {
@@ -139,7 +125,7 @@ private:
   std::chrono::milliseconds watchdog_megamiss_timeout_;
   std::chrono::milliseconds watchdog_kill_timeout_;
   std::chrono::milliseconds watchdog_multikill_timeout_;
-  RateLimitServiceConfigPtr ratelimit_service_config_;
+  Extensions::Filters::Common::RateLimit::RateLimitServiceConfigPtr ratelimit_service_config_;
 };
 
 /**
